worked on tools: normal anvil repair working replace placeholder code for actual behaviour use superclass when possible implemented shovels (bugged)

+++ src/main/java/org/oilmod/oilforge/rep/item/StandardItemStateFR.java
@@ -6,7 +6,7 @@ public class StandardItemStateFR extends ItemStateBase {
     }
 
     @Override
-    public short getNMS() {
+    public int getNMS() {
         return 0;
     }
